pyproject.toml
[project]
name = "groupme-meme-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=2.0.0",
    "fuzzywuzzy>=0.18.0",
    "groupyapi>=0.10.3",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
]


[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[dependency-groups]
dev = []
